chore: separate postList component folder
