+++ apps/minsu_dev/src/app/page.tsx
@@ -1,6 +1,6 @@
-import PostList from '@/components/common/post/list';
 import { createClient } from '@/utils/supabase/server';
 import { cookies } from 'next/headers';
+import { PostList } from '@/foo/home/list';
 import styles from '../foo/layout.module.css';
 
 export default async function Home() {
