+++ apps/minsu_dev/src/foo/home/list/index.ts
@@ -1,0 +1,1 @@
+export { default as PostList } from './list';
